Refactor abstractmethods
