+++ fault/tester/__init__.py
@@ -1,0 +1,2 @@
+from .staged_tester import Tester
+from .symbolic_tester import SymbolicTester
