+++ fault/tester/staged_tester.py
@@ -1,3 +1,4 @@
+from .abstract_tester import AbstractTester
 import fault
 import inspect
 import fault
@@ -18,10 +19,11 @@ from fault.verilogams_target import VerilogAMSTarget
 from fault.spice_target import SpiceTarget
 from fault.actions import Loop, While, If
 from fault.circuit_utils import check_interface_is_subset
-from fault.wrapper import CircuitWrapper, PortWrapper
+from fault.wrapper import PortWrapper
 from fault.file import File
 from fault.select_path import SelectPath
 from fault.wrapped_internal_port import WrappedVerilogInternalPort
+from ..magma_utils import is_recursive_type
 import fault.expression as expression
 import os
 import inspect
@@ -31,7 +33,7 @@ import tempfile
 from hwtypes import BitVector
 
 
-class Tester:
+class Tester(AbstractTester):
     """
     The fault `Tester` object provides a mechanism in Python to construct tests
     for magma circuits.  The `Tester` is instantiated with a specific magma
@@ -75,24 +77,8 @@ class Tester:
         `expect_strict_default`: if True, use strict equality check if
         not specified by the user.
         """
-        self._circuit = circuit
-        self.poke_delay_default = poke_delay_default
-        self.expect_strict_default = expect_strict_default
-        self.actions = []
-        if clock is not None and not isinstance(clock, m.Clock):
-            raise TypeError(f"Expected clock port: {clock, type(clock)}")
-        self.clock = clock
-        # Make sure the user has initialized the clock before stepping it
-        # While verilator initializes the clock value to 0, this assumption
-        # does not hold for system verilog, so we log a warning (as to not
-        # break existing TBs)
-        self.clock_initialized = False
-        # Only report once, in case the user calls step with an uninitialized
-        # clock many times
-        self.clock_init_warning_reported = False
-        if reset is not None and not isinstance(reset, m.Reset):
-            raise TypeError(f"Expected reset port: {reset, type(reset)}")
-        self.reset_port = reset
+        super().__init__(circuit, clock, reset, poke_delay_default,
+                         expect_strict_default)
         self.targets = {}
         # For public verilator modules
         self.verilator_includes = []
@@ -127,20 +113,6 @@ class Tester:
             return SpiceTarget(self._circuit, **kwargs)
         raise NotImplementedError(target)
 
-    def is_recursive_type(self, T):
-        return issubclass(T, m.Tuple) or issubclass(T, m.Array) and \
-            not issubclass(T.T, m.Digital)
-
-    def get_type(self, port):
-        if isinstance(port, SelectPath):
-            port = port[-1]
-        if isinstance(port, WrappedVerilogInternalPort):
-            type_ = port.type_
-            print(port.type_)
-        else:
-            type_ = type(port)
-        return type_
-
     def poke(self, port, value, delay=None):
         """
         Set `port` to be `value`
@@ -171,11 +143,11 @@ class Tester:
 
         # implement poke
         if isinstance(port, SelectPath):
-            if (self.is_recursive_type(type(port[-1]))
+            if (is_recursive_type(type(port[-1]))
                 or (not isinstance(port[-1], WrappedVerilogInternalPort) and
                     isinstance(port[-1].name, m.ref.AnonRef))):
                 return recurse(port[-1])
-        elif self.is_recursive_type(type(port)):
+        elif is_recursive_type(type(port)):
             return recurse(port)
         elif not isinstance(port, WrappedVerilogInternalPort) and\
                 isinstance(port.name, m.ref.AnonRef):
@@ -183,7 +155,7 @@ class Tester:
 
         if not isinstance(value, (LoopIndex, actions.FileRead,
                                   expression.Expression)):
-            type_ = self.get_type(port)
+            type_ = self.get_port_type(port)
             value = make_value(type_, value)
         self.actions.append(actions.Poke(port, value, delay=delay))
 
@@ -234,11 +206,11 @@ class Tester:
                                 **kwargs)
 
         if isinstance(port, SelectPath):
-            if (self.is_recursive_type(type(port[-1]))
+            if (is_recursive_type(type(port[-1]))
                 or (not isinstance(port[-1], WrappedVerilogInternalPort)
                     and isinstance(port[-1].name, m.ref.AnonRef))):
                 return recurse(port[-1])
-        elif self.is_recursive_type(type(port)):
+        elif is_recursive_type(type(port)):
             return recurse(port)
         elif not isinstance(port, WrappedVerilogInternalPort) and \
                 isinstance(port.name, m.ref.AnonRef):
@@ -247,7 +219,7 @@ class Tester:
         # implement expect
         if not isinstance(value, (actions.Peek, PortWrapper, actions.FileRead,
                                   LoopIndex, expression.Expression)):
-            type_ = self.get_type(port)
+            type_ = self.get_port_type(port)
             value = make_value(type_, value)
         self.actions.append(actions.Expect(port=port, value=value,
                                            strict=strict, caller=caller,
@@ -396,15 +368,6 @@ class Tester:
                               self.actions]
         return new_tester
 
-    def zero_inputs(self):
-        """
-        Set all the input ports to 0, useful for intiializing everything to a
-        known value
-        """
-        for name, port in self._circuit.IO.ports.items():
-            if port.is_input():
-                self.poke(self._circuit.interface.ports[name], 0)
-
     def clear(self):
         """
         Reset the tester by removing any existing actions. Useful for reusing a
@@ -425,10 +388,6 @@ class Tester:
 
     def verilator_include(self, module_name):
         self.verilator_includes.append(module_name)
-
-    @property
-    def circuit(self):
-        return CircuitWrapper(self._circuit, self)
 
     def loop(self, n_iter):
         """
@@ -486,75 +445,6 @@ class Tester:
         var = actions.Var(name, _type)
         self.actions.append(var)
         return var
-
-    def wait_on(self, cond):
-        loop = self._while(cond)
-        loop.step()
-
-    def wait_until_low(self, signal):
-        self.wait_on(self.peek(signal))
-
-    def wait_until_high(self, signal):
-        self.wait_on(~self.peek(signal))
-
-    def wait_until_negedge(self, signal):
-        self.wait_until_high(signal)
-        self.wait_until_low(signal)
-
-    def wait_until_posedge(self, signal, steps_per_iter=1):
-        self.wait_until_low(signal)
-        self.wait_until_high(signal)
-
-    def pulse_high(self, signal):
-        # first make sure the signal is actually low to begin with
-        self.expect(signal, 0)
-
-        # first set the signal high, then bring it low again
-        self.poke(signal, 1)
-        self.poke(signal, 0)
-
-    def pulse_low(self, signal):
-        # first make sure the signal is actually high to begin with
-        self.expect(signal, 1)
-
-        # first set the signal low, then bring it high again
-        self.poke(signal, 0)
-        self.poke(signal, 1)
-
-    def sync_reset(self, active_high=True, cycles=1):
-        # assert reset and set clock to zero
-        self.poke(self.reset_port, 1 if active_high else 0)
-        self.poke(self.clock, 0)
-
-        # wait the desired number of clock cycles
-        self.step(2 * cycles)
-
-        # de-assert reset
-        self.poke(self.reset_port, 0 if active_high else 1)
-
-    def internal(self, *args):
-        # return a SelectPath containing the desired path
-        return SelectPath([self.circuit] + list(args))
-
-    def __call__(self, *args, **kwargs):
-        """
-        Poke the inputs of the circuit using *args (ordered, anonymous input
-        reference, excluding clocks) and **kwargs.  **kwargs will overwrite any
-        inputs written by *args.
-
-        Evaluate the circuit
-
-        Return the "peeked" output(s) of the circuit (tuple for multiple
-        outputs)
-        """
-        for arg, port in zip(args, self._circuit.interface.outputs()):
-            self.poke(port, arg)
-        self.eval()
-        result = tuple(self.peek(port) for port in
-                       self._circuit.interface.inputs())
-        if len(result) == 1:
-            return result[0]
-        return result
 
 
 class LoopIndex:
